move csrmetrics into pkg.
To allow visibility from //pkg/...
