+++ cmd/gcp-controller-manager/BUILD
@@ -32,7 +32,8 @@ go_library(
         "//cmd/gcp-controller-manager:__subpackages__",
     ],
     deps = [
-        "//cmd/gcp-controller-manager/csrmetrics:go_default_library",
+        "//pkg/csrmetrics:go_default_library",
+        "//pkg/csrutil:go_default_library",
         "//pkg/nodeidentity:go_default_library",
         "//pkg/tpmattest:go_default_library",
         "//vendor/cloud.google.com/go/compute/metadata:go_default_library",
